Creación de nuevos directorios y refactorizado de código.
